+++ lib/go-tc/constants.go
@@ -41,6 +41,7 @@ const TenantUserNotAuthError = ErrorConstant("user not authorized for requested 
 const TenantDSUserNotAuthError = ErrorConstant("user not authorized for requested delivery service tenant")
 const NeedsAtLeastOneIPError = ErrorConstant("both IP and IP6 addresses are empty")
 const EmptyAddressCannotBeAServiceAddressError = ErrorConstant("an empty IP or IPv6 address cannot be marked as a service address")
+const NeedsAtLeastOneServiceAddressError = ErrorConstant("at least one address must be marked as a service address")
 
 // AlertLevel is used for specifying or comparing different levels of alerts.
 type AlertLevel int